Package visualisation .so files using auditwheel, without packaging cuda libraries
This is done by patching out the libcuda.so and libnvrtc.so dynamic dependencies, running auditwheel then re-adding the cuda dependencies.

This mostly works, but either the system installed libGLEW must by LD_PRELOADed, or LD_LIBRARY_PATH and LD_PRELOAD can be used to enable use of the packaged libGLEW.

Ideally this needs resolving, but it's unclear how currently, given libglew is related to the graphics driver.