Use audithweel repair and patchelf to pack .so files into the wheel
This is not ready for use, as it doesn't neccisarily produce vis wheels that actually work

Patchelf is used to skip libcuda.so etc from being packed, so wheels would not be compliant until that is resolved
But packing the vis .so's would be a step towards complaince.

There are still (potential) issues with opengl libraries being packed that leads to bad wheels.

In general, must use the old opengl ABI (libGL.so) as this is allowed as an external linkage
Rather than linking against libOpenGL.so.

libGLX and libGLDispatch which are also parts of the newer abi are implicitly requires still, by GLU and libGL, so fixing GLU may lead to portable, packaged wheels